+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.15
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.35.11
+
+## 0.1.14
+
+### Patch Changes
+
+-   Updated dependencies [69f62ec]
+    -   @sap-ux/fe-fpm-writer@0.35.10
+
+## 0.1.13
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.35.9
+
+## 0.1.12
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.35.8
+
 ## 0.1.11
 
 ### Patch Changes