+++ packages/adp-flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,116 @@
 # @sap-ux/adp-flp-config-sub-generator
 
+## 0.1.45
+
+### Patch Changes
+
+-   Updated dependencies [e53c1f8]
+    -   @sap-ux/adp-tooling@0.15.0
+    -   @sap-ux/flp-config-inquirer@0.3.45
+
+## 0.1.44
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.25
+-   @sap-ux/adp-tooling@0.14.44
+-   @sap-ux/flp-config-inquirer@0.3.44
+
+## 0.1.43
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.43
+-   @sap-ux/flp-config-inquirer@0.3.43
+
+## 0.1.42
+
+### Patch Changes
+
+-   Updated dependencies [d75db00]
+    -   @sap-ux/fiori-generator-shared@0.13.2
+    -   @sap-ux/flp-config-inquirer@0.3.42
+    -   @sap-ux/inquirer-common@0.7.24
+    -   @sap-ux/adp-tooling@0.14.42
+
+## 0.1.41
+
+### Patch Changes
+
+-   Updated dependencies [58cdce6]
+    -   @sap-ux/project-access@1.30.7
+    -   @sap-ux/adp-tooling@0.14.41
+    -   @sap-ux/axios-extension@1.22.3
+    -   @sap-ux/fiori-generator-shared@0.13.1
+    -   @sap-ux/flp-config-inquirer@0.3.41
+    -   @sap-ux/system-access@0.6.10
+    -   @sap-ux/inquirer-common@0.7.23
+
+## 0.1.40
+
+### Patch Changes
+
+-   Updated dependencies [58abe82]
+    -   @sap-ux/fiori-generator-shared@0.13.0
+    -   @sap-ux/flp-config-inquirer@0.3.40
+    -   @sap-ux/inquirer-common@0.7.22
+    -   @sap-ux/adp-tooling@0.14.40
+
+## 0.1.39
+
+### Patch Changes
+
+-   69f62ec: i18next upgrade to 25.3.0
+-   Updated dependencies [69f62ec]
+    -   @sap-ux/fiori-generator-shared@0.12.16
+    -   @sap-ux/flp-config-inquirer@0.3.39
+    -   @sap-ux/inquirer-common@0.7.21
+    -   @sap-ux/adp-tooling@0.14.39
+    -   @sap-ux/store@1.1.2
+    -   @sap-ux/system-access@0.6.10
+
+## 0.1.38
+
+### Patch Changes
+
+-   e9beeb4: fix: Various bugs in FLP config generator for Adaptation Project
+-   Updated dependencies [e9beeb4]
+    -   @sap-ux/adp-tooling@0.14.38
+    -   @sap-ux/flp-config-inquirer@0.3.38
+
+## 0.1.37
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.37
+-   @sap-ux/project-access@1.30.6
+-   @sap-ux/flp-config-inquirer@0.3.37
+-   @sap-ux/axios-extension@1.22.3
+-   @sap-ux/fiori-generator-shared@0.12.15
+-   @sap-ux/system-access@0.6.9
+-   @sap-ux/inquirer-common@0.7.20
+
+## 0.1.36
+
+### Patch Changes
+
+-   Updated dependencies [7a4543e]
+    -   @sap-ux/project-access@1.30.5
+    -   @sap-ux/adp-tooling@0.14.36
+    -   @sap-ux/axios-extension@1.22.3
+    -   @sap-ux/fiori-generator-shared@0.12.14
+    -   @sap-ux/flp-config-inquirer@0.3.36
+    -   @sap-ux/system-access@0.6.9
+    -   @sap-ux/inquirer-common@0.7.19
+
+## 0.1.35
+
+### Patch Changes
+
+-   Updated dependencies [6c2d08a]
+    -   @sap-ux/adp-tooling@0.14.35
+    -   @sap-ux/flp-config-inquirer@0.3.35
+
 ## 0.1.34
 
 ### Patch Changes