+++ packages/create/CHANGELOG.md
@@ -1,5 +1,194 @@
 # @sap-ux/create
 
+## 0.13.74
+
+### Patch Changes
+
+-   Updated dependencies [e53c1f8]
+    -   @sap-ux/abap-deploy-config-writer@0.2.0
+    -   @sap-ux/adp-tooling@0.15.0
+    -   @sap-ux/flp-config-inquirer@0.3.45
+    -   @sap-ux/preview-middleware@0.20.62
+    -   @sap-ux/app-config-writer@0.6.23
+
+## 0.13.73
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.35
+-   @sap-ux/adp-tooling@0.14.44
+-   @sap-ux/flp-config-inquirer@0.3.44
+-   @sap-ux/preview-middleware@0.20.61
+-   @sap-ux/app-config-writer@0.6.23
+
+## 0.13.72
+
+### Patch Changes
+
+-   Updated dependencies [eafb486]
+    -   @sap-ux/preview-middleware@0.20.60
+    -   @sap-ux/app-config-writer@0.6.23
+
+## 0.13.71
+
+### Patch Changes
+
+-   Updated dependencies [e5de360]
+    -   @sap-ux/preview-middleware@0.20.59
+    -   @sap-ux/app-config-writer@0.6.23
+
+## 0.13.70
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.43
+-   @sap-ux/flp-config-inquirer@0.3.43
+-   @sap-ux/preview-middleware@0.20.58
+-   @sap-ux/app-config-writer@0.6.23
+
+## 0.13.69
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.34
+-   @sap-ux/cap-config-writer@0.10.20
+-   @sap-ux/flp-config-inquirer@0.3.42
+-   @sap-ux/adp-tooling@0.14.42
+-   @sap-ux/preview-middleware@0.20.57
+-   @sap-ux/app-config-writer@0.6.23
+
+## 0.13.68
+
+### Patch Changes
+
+-   Updated dependencies [58cdce6]
+    -   @sap-ux/project-access@1.30.7
+    -   @sap-ux/abap-deploy-config-writer@0.1.17
+    -   @sap-ux/adp-tooling@0.14.41
+    -   @sap-ux/app-config-writer@0.6.23
+    -   @sap-ux/cap-config-writer@0.10.19
+    -   @sap-ux/flp-config-inquirer@0.3.41
+    -   @sap-ux/mockserver-config-writer@0.9.8
+    -   @sap-ux/odata-service-writer@0.27.12
+    -   @sap-ux/preview-middleware@0.20.56
+    -   @sap-ux/system-access@0.6.10
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.33
+
+## 0.13.67
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.32
+-   @sap-ux/cap-config-writer@0.10.18
+-   @sap-ux/flp-config-inquirer@0.3.40
+-   @sap-ux/adp-tooling@0.14.40
+-   @sap-ux/preview-middleware@0.20.55
+-   @sap-ux/app-config-writer@0.6.22
+
+## 0.13.66
+
+### Patch Changes
+
+-   Updated dependencies [69f62ec]
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.31
+    -   @sap-ux/mockserver-config-writer@0.9.7
+    -   @sap-ux/odata-service-writer@0.27.11
+    -   @sap-ux/flp-config-inquirer@0.3.39
+    -   @sap-ux/app-config-writer@0.6.22
+    -   @sap-ux/cap-config-writer@0.10.17
+    -   @sap-ux/adp-tooling@0.14.39
+    -   @sap-ux/preview-middleware@0.20.54
+    -   @sap-ux/system-access@0.6.10
+    -   @sap-ux/ui5-config@0.29.0
+    -   @sap-ux/abap-deploy-config-writer@0.1.16
+
+## 0.13.65
+
+### Patch Changes
+
+-   fca048c: Remove unused dependency to @sap-ux/cards-editor-config-writer
+
+## 0.13.64
+
+### Patch Changes
+
+-   e9beeb4: fix: Various bugs in FLP config generator for Adaptation Project
+-   Updated dependencies [e9beeb4]
+    -   @sap-ux/adp-tooling@0.14.38
+    -   @sap-ux/flp-config-inquirer@0.3.38
+    -   @sap-ux/preview-middleware@0.20.53
+    -   @sap-ux/app-config-writer@0.6.21
+
+## 0.13.63
+
+### Patch Changes
+
+-   Updated dependencies [c0fa1d1]
+    -   @sap-ux/ui5-config@0.29.0
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.30
+    -   @sap-ux/abap-deploy-config-writer@0.1.15
+    -   @sap-ux/adp-tooling@0.14.37
+    -   @sap-ux/app-config-writer@0.6.21
+    -   @sap-ux/cards-editor-config-writer@0.5.4
+    -   @sap-ux/mockserver-config-writer@0.9.6
+    -   @sap-ux/odata-service-writer@0.27.10
+    -   @sap-ux/project-access@1.30.6
+    -   @sap-ux/flp-config-inquirer@0.3.37
+    -   @sap-ux/preview-middleware@0.20.52
+    -   @sap-ux/cap-config-writer@0.10.16
+    -   @sap-ux/system-access@0.6.9
+
+## 0.13.62
+
+### Patch Changes
+
+-   Updated dependencies [7a4543e]
+    -   @sap-ux/project-access@1.30.5
+    -   @sap-ux/abap-deploy-config-writer@0.1.14
+    -   @sap-ux/adp-tooling@0.14.36
+    -   @sap-ux/app-config-writer@0.6.20
+    -   @sap-ux/cap-config-writer@0.10.15
+    -   @sap-ux/cards-editor-config-writer@0.5.3
+    -   @sap-ux/flp-config-inquirer@0.3.36
+    -   @sap-ux/mockserver-config-writer@0.9.5
+    -   @sap-ux/odata-service-writer@0.27.9
+    -   @sap-ux/preview-middleware@0.20.51
+    -   @sap-ux/system-access@0.6.9
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.29
+
+## 0.13.61
+
+### Patch Changes
+
+-   Updated dependencies [fc8cc4a]
+    -   @sap-ux/preview-middleware@0.20.50
+    -   @sap-ux/app-config-writer@0.6.19
+
+## 0.13.60
+
+### Patch Changes
+
+-   Updated dependencies [67cf59e]
+    -   @sap-ux/preview-middleware@0.20.49
+    -   @sap-ux/app-config-writer@0.6.19
+
+## 0.13.59
+
+### Patch Changes
+
+-   Updated dependencies [6c2d08a]
+    -   @sap-ux/adp-tooling@0.14.35
+    -   @sap-ux/preview-middleware@0.20.48
+    -   @sap-ux/flp-config-inquirer@0.3.35
+    -   @sap-ux/app-config-writer@0.6.19
+
+## 0.13.58
+
+### Patch Changes
+
+-   Updated dependencies [a9df80c]
+    -   @sap-ux/app-config-writer@0.6.19
+
 ## 0.13.57
 
 ### Patch Changes