+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.13.2
+
+### Patch Changes
+
+-   d75db00: Append leading forward slash
+
+## 0.13.1
+
+### Patch Changes
+
+-   Updated dependencies [58cdce6]
+    -   @sap-ux/project-access@1.30.7
+    -   @sap-ux/telemetry@0.6.10
+
+## 0.13.0
+
+### Minor Changes
+
+-   58abe82: Add optional isActive parameter to isExtensionInstalled function & add support for checking command registration
+
+## 0.12.16
+
+### Patch Changes
+
+-   69f62ec: i18next upgrade to 25.3.0
+    -   @sap-ux/telemetry@0.6.9
+
+## 0.12.15
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.6
+-   @sap-ux/telemetry@0.6.8
+
+## 0.12.14
+
+### Patch Changes
+
+-   Updated dependencies [7a4543e]
+    -   @sap-ux/project-access@1.30.5
+    -   @sap-ux/telemetry@0.6.7
+
 ## 0.12.13
 
 ### Patch Changes