+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @sap-ux/odata-service-writer
 
+## 0.27.12
+
+### Patch Changes
+
+-   Updated dependencies [58cdce6]
+    -   @sap-ux/project-access@1.30.7
+    -   @sap-ux/mockserver-config-writer@0.9.8
+
+## 0.27.11
+
+### Patch Changes
+
+-   69f62ec: i18next upgrade to 25.3.0
+-   Updated dependencies [69f62ec]
+    -   @sap-ux/mockserver-config-writer@0.9.7
+    -   @sap-ux/ui5-config@0.29.0
+
+## 0.27.10
+
+### Patch Changes
+
+-   Updated dependencies [c0fa1d1]
+    -   @sap-ux/ui5-config@0.29.0
+    -   @sap-ux/mockserver-config-writer@0.9.6
+    -   @sap-ux/project-access@1.30.6
+
+## 0.27.9
+
+### Patch Changes
+
+-   Updated dependencies [7a4543e]
+    -   @sap-ux/project-access@1.30.5
+    -   @sap-ux/mockserver-config-writer@0.9.5
+
 ## 0.27.8
 
 ### Patch Changes