+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sap-ux/telemetry
 
+## 0.6.10
+
+### Patch Changes
+
+-   Updated dependencies [58cdce6]
+    -   @sap-ux/project-access@1.30.7
+
+## 0.6.9
+
+### Patch Changes
+
+-   Updated dependencies [69f62ec]
+    -   @sap-ux/store@1.1.2
+    -   @sap-ux/ui5-config@0.29.0
+
+## 0.6.8
+
+### Patch Changes
+
+-   Updated dependencies [c0fa1d1]
+    -   @sap-ux/ui5-config@0.29.0
+    -   @sap-ux/project-access@1.30.6
+
+## 0.6.7
+
+### Patch Changes
+
+-   Updated dependencies [7a4543e]
+    -   @sap-ux/project-access@1.30.5
+
 ## 0.6.6
 
 ### Patch Changes