+++ packages/ui5-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/ui5-proxy-middleware
 
+## 1.5.7
+
+### Patch Changes
+
+-   69f62ec: i18next upgrade to 25.3.0
+    -   @sap-ux/ui5-config@0.29.0
+
+## 1.5.6
+
+### Patch Changes
+
+-   Updated dependencies [c0fa1d1]
+    -   @sap-ux/ui5-config@0.29.0
+
+## 1.5.5
+
+### Patch Changes
+
+-   7a4543e: fix: remove usage of static webapp folder
+
 ## 1.5.4
 
 ### Patch Changes